Refactor and create dplaapi.queries
Create dplaapi.queries subpackage to better organize search_query and
mlt_query, in advance of making SearchQuery and MLTQuery descendents of
a base class.
